Rearranged the application to a feature-sliced. Decided to have just a regular modal for a profile. Added a couple of custom hooks, and also rearranged them to a feature-sliced
