+++ frontend/src/entities/chat/ui/chat-components/chat-input/ChatInput.tsx
