+++ frontend/src/entities/chat/ui/chat-components/message-bubble/model/types.ts
