+++ frontend/src/entities/chat/ui/chat-components/message-bubble/ui/MessageBubble.tsx
