+++ frontend/src/entities/user/model/useUser.ts
@@ -1,0 +1,56 @@
+import { useNavigate } from "react-router";
+import { useUserStore } from "../../../util/model/store/zustand";
+import { authPort, port } from "../../../util/ui/ProtectedRoutes";
+
+export const useUser = () => {
+  const setUser = useUserStore((state) => state.setUser);
+  const navigate = useNavigate();
+  const checkUser = async () => {
+    try {
+      const response = await fetch(`${port}/me`, {
+        method: "GET",
+        credentials: "include",
+      });
+
+      if (!response.ok) {
+        const isOk = await refresh();
+        if (!isOk) {
+          setUser(undefined);
+        }
+      } else {
+        const { email, isVerified, login, role, id } = await response.json();
+        if (isVerified === false && email) {
+          navigate(`/verify?email=${email}`);
+        }
+        if (isVerified) {
+          setUser({ email, login, role, id });
+        }
+      }
+    } catch (e) {
+      setUser(undefined);
+      console.log(e);
+    } finally {
+      return true;
+    }
+  };
+
+  const refresh = async () => {
+    const refreshResponse = await fetch(`${authPort}/api/refresh`, {
+      method: "GET",
+      credentials: "include",
+    });
+    if (refreshResponse.ok) {
+      const { email, isVerified, login, role, id } =
+        await refreshResponse.json();
+      if (isVerified) {
+        setUser({ email, login, role, id });
+      }
+    } else {
+      setUser(undefined);
+      return false;
+    }
+    return true;
+  };
+
+  return { checkUser };
+};
