+++ frontend/src/features/authentication/ui/auth-input/AuthInput.tsx
