+++ frontend/src/routes/ProfilePage.tsx
@@ -1,0 +1,5 @@
+export default function ProfilePage(){
+  return (<main className="ml-[50px] h-screen bg-[#18191A] w-[calc(100%-50px)] flex flex-row">
+    
+  </main>
+)}
