--- frontend/src/store/store.ts
@@ -1,10 +1,0 @@
-import { configureStore } from "@reduxjs/toolkit";
-import userReducer from "./userReducer"
-export const store = configureStore({
-  reducer: {
-    user: userReducer,
-  }
-})
-
-export type RootState = ReturnType<typeof store.getState>
-export type AppDispatch = typeof store.dispatch
