+++ frontend/src/util/model/socket/socket.ts
@@ -1,0 +1,8 @@
+import { io } from "socket.io-client";
+import { port } from "../../ui/ProtectedRoutes";
+export const socket = io(port, {
+  path: "/socket.io",
+  withCredentials: true,
+  transports: ["websocket"],
+  autoConnect: true,
+});
